auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]==1.7.4
python-jose
pydantic
email-validator
bcrypt==4.0.1